requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-groq==1.1.0
langchain-huggingface==1.1.0
langchain-text-splitters==1.0.0
python-dotenv==1.2.1
python-multipart